FRUC starts with average now

+++ errorsVideos.m
@@ -4,8 +4,8 @@ totalMse=0;
 totalPsnr = 0;
 for i=1:1:max(size(mask))
     if mask(i)==1
-    totalMse = totalMse + mask(i) * immse(origin(:,:,i),restored(:,:,i));
-    totalPsnr = totalPsnr + mask(i) * psnr(origin(:,:,i),restored(:,:,i),255);
+    totalMse = totalMse + mask(i) * immse(uint8(origin(:,:,i)),uint8(restored(:,:,i)));
+    totalPsnr = totalPsnr + mask(i) * psnr(uint8(origin(:,:,i)),uint8(restored(:,:,i)),255);
     end
 end
 mse = double(totalMse)/double(sum(mask));
